call_home: make reverse ssh port dynamic
This is not ideal, but it does seem to be a bit better, because if the port
is fixed, the device may have difficulties reconnecting back, or take
longer.
This way at least if a device fails to connect back on a specific port, it
tries the next.

This also removes one manual step in setting up a jig, where we need to
assign manually ports to each test jig.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>