lib/production.sh: display firmware version at start of loop
This change re-uses some of the logic in the `lib/flash.sh` script.
The idea is that we should check for all FW files to be present before
starting a loop.

And also to display the FW version that will be flashed to make sure that
we are flashing the version we expect to flash.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>